% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb.R
\name{adapt_data_table_format}
\alias{adapt_data_table_format}
\title{From non-SQL DynamoDB table to spread tibble}
\usage{
adapt_data_table_format(table, tzone = "Europe/Paris")
}
\arguments{
\item{table}{data.frame, response from DynamoDB query}

\item{tzone}{character, time zone}
}
\value{
tibble
}
\description{
From non-SQL DynamoDB table to spread tibble
}
