% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{adapt_yearly_timeseries}
\alias{adapt_yearly_timeseries}
\title{Adapt an yearly time series dataframe to timezone and year}
\usage{
adapt_yearly_timeseries(df, tz_out = NULL, year_out = NULL)
}
\arguments{
\item{df}{tibble with first column being \code{datetime}}

\item{tz_out}{character, time zone of the desired \code{datetime}}

\item{year_out}{integer, year of the desired \code{datetime}}
}
\value{
tibble
}
\description{
Adapt an yearly time series dataframe to timezone and year
}
