% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{aggregate_timeseries}
\alias{aggregate_timeseries}
\title{Aggregate multiple timeseries columns to a single one}
\usage{
aggregate_timeseries(df, varname)
}
\arguments{
\item{df}{data.frame or tibble, first column of name \code{datetime} being of class datetime and rest of columns being numeric}

\item{varname}{character, name of the aggregation column}
}
\value{
tibble
}
\description{
The first column \code{datetime} will be kept.
}
