% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseSelectorModule.R
\name{caseSelector}
\alias{caseSelector}
\title{Server function of caseSelector module}
\usage{
caseSelector(id, df1, df2, plot_func, plot_conf, ...)
}
\arguments{
\item{id}{character, module ID}

\item{df1}{data.frame or tibble, data of case 1 being \code{datetime} the first column}

\item{df2}{data.frame or tibble, data of case 2 being \code{datetime} the first column}

\item{plot_func}{function to plot the dygraph output, being the first parameter the configuration tibble}

\item{plot_conf}{tibble of the plot configuration}

\item{...}{extra parameters passed to dyOptions}
}
\value{
reactive \code{case} module input
}
\description{
Server function of caseSelector module
}
