% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseSelectorModule.R
\name{caseSelectorUI}
\alias{caseSelectorUI}
\title{UI function for caseSelector module}
\usage{
caseSelectorUI(id, tabs, height)
}
\arguments{
\item{id}{character, module ID}

\item{tabs}{character vector, tabs names}

\item{height}{Dygraphs height. Must be a valid CSS unit (like "100\%", "400px", "auto") or a number, which will be coerced to a string and have "px" appended.}
}
\description{
UI function for caseSelector module
}
