% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{change_timeseries_year}
\alias{change_timeseries_year}
\title{Change the year of a time series data frame keeping the original weekdays}
\usage{
change_timeseries_year(df, year_out)
}
\arguments{
\item{df}{tibble with first column being \code{datetime}}

\item{year_out}{character, year of the desired \code{datetime}}
}
\value{
tibble
}
\description{
The input \code{df} must contain full-week time-series profiles in order to
arrange the data according to the day of the week. For example, if the first
day in the \code{df} is a Monday the last one must be a Sunday.
}
