% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{compare_kpis}
\alias{compare_kpis}
\title{Compare KPIs (final kpis - original kpis)}
\usage{
compare_kpis(kpis_o, kpis_f)
}
\arguments{
\item{kpis_o}{named list with original kpis values}

\item{kpis_f}{named list with final kpis values}
}
\value{
named list
}
\description{
Compare KPIs (final kpis - original kpis)
}
