% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{date_to_datetime_with_tz}
\alias{date_to_datetime_with_tz}
\title{Convert date to datetime with a timezone}
\usage{
date_to_datetime_with_tz(date, tzone = "UTC")
}
\arguments{
\item{date}{date}

\item{tzone}{character, time zone}
}
\value{
datetime
}
\description{
Only valid for positive time zone differences (right part pf the mapamundi)
}
