% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{date_to_timestamp}
\alias{date_to_timestamp}
\title{Convert date or datetime value to timestamp number}
\usage{
date_to_timestamp(date, tzone = "Europe/Paris", milliseconds = T)
}
\arguments{
\item{date}{date or datetime value}

\item{tzone}{character, time-zone of the current time}

\item{milliseconds}{logical, whether the timestamp is in milliseconds or seconds}
}
\value{
numeric
}
\description{
Convert date or datetime value to timestamp number
}
