% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{decrease_numeric_resolution}
\alias{decrease_numeric_resolution}
\title{Decrease resolution of a numeric vector}
\usage{
decrease_numeric_resolution(y, n, method = c("average", "first"))
}
\arguments{
\item{y}{original numeric vector}

\item{n}{integer, number of intra-values (counting the original value as the first one)}

\item{method}{character, being \code{average} or \code{first} as valid options}
}
\value{
numeric vector
}
\description{
Decrease resolution of a numeric vector
}
