% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb.R
\name{delete_df_from_dynamodb}
\alias{delete_df_from_dynamodb}
\title{Delete data frame from DynamoDB table}
\usage{
delete_df_from_dynamodb(dynamodb_table, df)
}
\arguments{
\item{dynamodb_table}{boto3 DynamoDB Table object}

\item{df}{data frame to put}
}
\description{
Delete data frame from DynamoDB table
}
