% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_detection.R
\name{detect_step_demand}
\alias{detect_step_demand}
\title{Title}
\usage{
detect_step_demand(df, colname, power, enable_half = FALSE)
}
\arguments{
\item{df}{data.frame or tibble, first column of name \code{datetime} being of class datetime and rest of columns being numeric}

\item{colname}{character, column name where the step load should be detected}

\item{power}{numeric, value of power (units according to \code{df}) of the step load}

\item{enable_half}{logical, whether to consider half of power demand during indices before and after the step into the final step profile.}
}
\value{
data.frame or tibble
}
\description{
Title
}
