% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{df_to_ts}
\alias{df_to_ts}
\title{Convert time-series data.frame to time-series object}
\usage{
df_to_ts(df)
}
\arguments{
\item{df}{data.frame or tibble, first column of name \code{datetime} being of class datetime and rest of columns being numeric}
}
\value{
xts time-series object
}
\description{
Convert time-series data.frame to time-series object
}
