% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_paws.R
\name{dynamodb_delete_items}
\alias{dynamodb_delete_items}
\title{Delete items from DynamoDB table}
\usage{
dynamodb_delete_items(
  dynamodb_obj,
  table_name,
  partition_key_name,
  partition_key_values,
  sorting_key_name,
  sorting_key_start,
  sorting_key_end
)
}
\arguments{
\item{dynamodb_obj}{DynamoDB object obtained with \code{get_dynamodb} function}

\item{table_name}{character, DynamoDB table name}

\item{partition_key_name}{character, name of the partiton key of the DynamoDB Table. It has to be a string variable}

\item{partition_key_values}{vector of values of the partiton key of the DynamoDB Table}

\item{sorting_key_name}{character, name of the sorting key of the DynamoDB Table. It has to be a numeric variable}

\item{sorting_key_start}{numeric, start value of the sorting key}

\item{sorting_key_end}{numeric, end value of the sorting key}
}
\description{
The table must have partition and sorting key and the items deleted are within a range of the sorting key
}
