% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb.R
\name{dynamodb_items_to_tbl}
\alias{dynamodb_items_to_tbl}
\title{Convert DynamoDB Items list to an R tibble}
\usage{
dynamodb_items_to_tbl(items)
}
\arguments{
\item{items}{list, Items from DynamoDB, being each Item an individual list of Python objects of class decimal.Decimal, character or logical. For other types a NA value will be returned.}
}
\value{
tibble
}
\description{
Convert DynamoDB Items list to an R tibble
}
