% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_python.R
\name{dynamodb_put_df_py}
\alias{dynamodb_put_df_py}
\title{Put data frame to DynamoDB table}
\usage{
dynamodb_put_df_py(dynamo_table, df)
}
\arguments{
\item{dynamo_table}{boto3 DynamoDB Table obtained with \code{get_dynamo_table} function}

\item{df}{data frame to put}
}
\description{
Put data frame to DynamoDB table
}
