% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_paws.R
\name{dynamodb_put_items}
\alias{dynamodb_put_items}
\title{Put items to DynamoDB table}
\usage{
dynamodb_put_items(dynamodb_obj, table_name, items)
}
\arguments{
\item{dynamodb_obj}{DynamoDB object obtained with \code{get_dynamodb} function}

\item{table_name}{character, DynamoDB table name}

\item{items}{list of value-key lists, with all Keys (partition and sorting) of the table and other optional attributes.
For example one item could be: \code{list(id = list(S = 'id123'), timestamp = list(N = 1622190385), data = list(temperature = list(N = 24.2), humidity = list(N = 32)))}}
}
\description{
Put items to DynamoDB table
}
