% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb.R
\name{dynamodb_query_items}
\alias{dynamodb_query_items}
\title{Title}
\usage{
dynamodb_query_items(
  dynamodb_table,
  partition_key_name,
  partition_key_values,
  sort_key_name = NULL,
  sort_key_values = NULL
)
}
\arguments{
\item{dynamodb_table}{boto3 DynamoDB Table object}

\item{partition_key_name}{character, name of the partiton key of the DynamoDB Table}

\item{partition_key_values}{vector of values of the partiton key of the DynamoDB Table}

\item{sort_key_name}{character, name of the sorting key of the DynamoDB Table}

\item{sort_key_values}{vector of values of the sorting key of the DynamoDB Table}
}
\value{
tibble
}
\description{
Title
}
