% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dyplot}
\alias{dyplot}
\title{Plot a timeseries data.frame in a Dygraphs HTML plot}
\usage{
dyplot(df, format = TRUE, css_file = NULL, ...)
}
\arguments{
\item{df}{data.frame or tibble, first column of name \code{datetime} being of class datetime and rest of columns being numeric}

\item{format}{logical, whether to format dygraph with custom CSS file}

\item{css_file}{character path to a CSS file to format dygraph plot. If NULL, custom CSS is applied to dygraph. Only used when \code{format} is \code{TRUE}.}

\item{...}{extra arguments to pass to \code{dygraphs::dyOptions} function. Only used when \code{format} is \code{TRUE}}
}
\value{
dygraph
}
\description{
First column of the data.frame must be a datetime or date column.
The rest of columns must be numeric of the same units.
}
