% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{fill_datetime}
\alias{fill_datetime}
\title{Fill NA values of a datetime sequence vector}
\usage{
fill_datetime(dttm)
}
\arguments{
\item{dttm}{datetime sequence vector}
}
\value{
filled datetime sequence vector
}
\description{
Fill NA values of a datetime sequence vector
}
