% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{fill_down_until}
\alias{fill_down_until}
\title{Fill down tibble columns until a maximum number of time slots}
\usage{
fill_down_until(data, vars, max_timeslots = 1)
}
\arguments{
\item{data}{data.frame or tibble}

\item{vars}{character, names of columns to fill}

\item{max_timeslots}{integer, maximum number of time slots to fill}
}
\value{
tibble
}
\description{
Fill down tibble columns until a maximum number of time slots
}
