% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{fill_from_past}
\alias{fill_from_past}
\title{Fill from past values}
\usage{
fill_from_past(tbl, colnames, back = 24)
}
\arguments{
\item{tbl}{tibble or data.frame, with NA values in some columns}

\item{colnames}{character or vector of characters, column names with NA values}

\item{back}{integer, number of positions (rows) to go back and get the filling value}
}
\value{
tibble or data.frame
}
\description{
Fill from past values
}
