% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericInputsModule.R
\name{get_component_value}
\alias{get_component_value}
\title{Get single component value}
\usage{
get_component_value(id, config_value, numericInputs_module)
}
\arguments{
\item{id}{character, input ID}

\item{config_value}{numeric, initial value}

\item{numericInputs_module}{numericInputServer object of the module}
}
\value{
numeric input value
}
\description{
Get single component value
}
