% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericInputsModule.R
\name{get_components_values}
\alias{get_components_values}
\title{Get components values from numericInput module}
\usage{
get_components_values(inputs_tbl, numericInputs_module)
}
\arguments{
\item{inputs_tbl}{data.frame or tibble, with inputs configuration}

\item{numericInputs_module}{numericInputServer object of the module}
}
\value{
named list with input values
}
\description{
Get components values from numericInput module
}
