% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_current_date}
\alias{get_current_date}
\title{Get current date from Internet, not from local machine}
\usage{
get_current_date()
}
\value{
date
}
\description{
Get current date from Internet, not from local machine
}
