% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_datetime_seq}
\alias{get_datetime_seq}
\title{Datetime sequence}
\usage{
get_datetime_seq(
  year,
  tzone,
  resolution_mins,
  fullyear = TRUE,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{year}{integer, year of the datetime sequence}

\item{tzone}{character, time-zone of the datetime sequence}

\item{resolution_mins}{integer, interval of minutes between two consecutive datetime values}

\item{fullyear}{boolean, whether to return a full-year sequence}

\item{start_date}{Date, if \code{fullyear} is \code{FALSE} set a starting date. Ignored when \code{fullyear} is \code{TRUE}.}

\item{end_date}{Date, if \code{fullyear} is \code{FALSE} set a final date. Ignored when \code{fullyear} is \code{TRUE}.}
}
\value{
vector of datetime values
}
\description{
Datetime sequence
}
