% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb.R
\name{get_dynamo_table}
\alias{get_dynamo_table}
\title{Get boto3 DynamoDB Table object}
\usage{
get_dynamo_table(dynamodb, table_name)
}
\arguments{
\item{dynamodb}{boto3 DynamoDB resource obtained with \code{get_dynamodb} function}

\item{table_name}{character, table name}
}
\value{
boto3 DynamoDB Table object
}
\description{
Get boto3 DynamoDB Table object
}
