% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb.R
\name{get_dynamodb}
\alias{get_dynamodb}
\title{Get DynamoDB paws resource}
\usage{
get_dynamodb(aws_access_key_id, aws_secret_access_key, region_name)
}
\arguments{
\item{aws_access_key_id}{AWS access key ID of a user with permissions for DynamoDB}

\item{aws_secret_access_key}{AWS secret access key of a user with permissions for DynamoDB}

\item{region_name}{AWS DynamoDB region name (e.g. eu-west-1)}
}
\value{
paws DynamoDB object list
}
\description{
Get DynamoDB paws resource
}
\details{
To create a user with permissions go to IAM AWS service and create a new user with Programmatic access, attaching existing policies directly for a fast configuration.
}
