% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb.R
\name{get_dynamodb_table}
\alias{get_dynamodb_table}
\title{Get boto3 DynamoDB Table object}
\usage{
get_dynamodb_table(dynamodb, table_name)
}
\arguments{
\item{dynamodb}{boto3 DynamoDB object}

\item{table_name}{character, table name}
}
\value{
boto3 DynamoDB Table object
}
\description{
Get boto3 DynamoDB Table object
}
