% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{get_energy_kpis}
\alias{get_energy_kpis}
\title{Obtain a list of energy indicators given the energy flows}
\usage{
get_energy_kpis(df, kg_co2_kwh = 0.5)
}
\arguments{
\item{df}{data.frame or tibble, with columns \code{datetime}, \code{consumption}, \code{production} and \code{kg_co2_kwh}}

\item{kg_co2_kwh}{factor of CO2 kg emissions per kWh of energy consumed from the distribution grid}
}
\value{
named list
}
\description{
Obtain a list of energy indicators given the energy flows
}
