% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb.R
\name{get_id_table}
\alias{get_id_table}
\title{Get data from a specific partition key}
\usage{
get_id_table(
  dynamodb,
  table_name,
  id,
  start_date,
  end_date,
  tzone = "Europe/Paris"
)
}
\arguments{
\item{dynamodb}{boto3 DynamoDB resource obtained with \code{get_dynamodb} function}

\item{table_name}{character, table name}

\item{id}{character, partition key ID}

\item{start_date}{Date, start date}

\item{end_date}{Date, end date}

\item{tzone}{character, time zone}
}
\value{
tibble
}
\description{
Get data from a specific partition key
}
