% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericInputsModule.R
\name{get_input_valid_value}
\alias{get_input_valid_value}
\title{Get single input value}
\usage{
get_input_valid_value(id, config_value, inputs_list)
}
\arguments{
\item{id}{character, input ID}

\item{config_value}{numeric, initial value}

\item{inputs_list}{\code{input} object of the module}
}
\value{
numeric input value
}
\description{
If the component is not in the inouts list, then the returned value is the
configuration value.
}
