% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericInputsModule.R
\name{get_inputs_valid_values}
\alias{get_inputs_valid_values}
\title{Get numeric inputs values from numericInput module}
\usage{
get_inputs_valid_values(inputs_conf, inputs_list)
}
\arguments{
\item{inputs_conf}{tibble, numericInputs parameters table with columns
\code{input} (numericInput ID), \code{label}, \code{value}, \code{min}, \code{max}, \code{step} and \code{show}}

\item{inputs_list}{\code{input} object of the module}
}
\value{
named list with input values
}
\description{
Get numeric inputs values from numericInput module
}
