% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{get_monthly_energy_kpis}
\alias{get_monthly_energy_kpis}
\title{For each month independently, obtain a list of energy indicators given the energy flows}
\usage{
get_monthly_energy_kpis(df)
}
\arguments{
\item{df}{data.frame or tibble, with columns \code{datetime}, \code{consumption}, \code{production} and \code{kg_co2_kwh}}
}
\value{
named list
}
\description{
For each month independently, obtain a list of energy indicators given the energy flows
}
