% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_percentage}
\alias{get_percentage}
\title{Get percentage of tibble variables}
\usage{
get_percentage(
  tbl,
  from,
  discard = NULL,
  percent = TRUE,
  keep = FALSE,
  digits = 2
)
}
\arguments{
\item{tbl}{tibble or data.frame}

\item{from}{character, column name of the base variable}

\item{discard}{character or character vector, columns to discard of the percentage operations (but included in the results)}

\item{percent}{logical, whether to output the percentages in percent or not}

\item{keep}{logical, whether to include the base variable in \code{from} argument to the output tibble}

\item{digits}{integer, number of digits to round the percentages}
}
\value{
tibble
}
\description{
The percentage is relative to the variable in \code{from} argument
}
