% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_time_resolution}
\alias{get_time_resolution}
\title{Return the time resolution of a datetime sequence}
\usage{
get_time_resolution(dttm_seq, units = "mins")
}
\arguments{
\item{dttm_seq}{datetime sequence}

\item{units}{character being one of "auto", "secs", "mins", "hours", "days" and "weeks"}
}
\value{
numeric
}
\description{
Return the time resolution of a datetime sequence
}
