% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvgis_api.R
\name{get_tmy}
\alias{get_tmy}
\title{Get TMY from PVGIS given the latitude and longitude coordinates.
For more information visit https://ec.europa.eu/jrc/en/pvgis}
\usage{
get_tmy(lat = 52.37, lon = 4.908)
}
\arguments{
\item{lat}{numeric, latitude coordinate}

\item{lon}{numeric, longitude coordinate}
}
\value{
tibble
}
\description{
Get TMY from PVGIS given the latitude and longitude coordinates.
For more information visit https://ec.europa.eu/jrc/en/pvgis
}
