% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_week_from_datetime}
\alias{get_week_from_datetime}
\title{Week date from datetime value}
\usage{
get_week_from_datetime(dttm)
}
\arguments{
\item{dttm}{datetime vector}
}
\value{
date vector
}
\description{
Week date from datetime value
}
