% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_week_total}
\alias{get_week_total}
\title{Summarise dataframe with weekly total column values}
\usage{
get_week_total(df)
}
\arguments{
\item{df}{data.frame or tibble, first column of name \code{datetime} being of class datetime and rest of columns being numeric}
}
\value{
tibble
}
\description{
Summarise dataframe with weekly total column values
}
