% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_weektime_from_datetime}
\alias{get_weektime_from_datetime}
\title{Week datetime from datetime value}
\usage{
get_weektime_from_datetime(dttm)
}
\arguments{
\item{dttm}{datetime vector}
}
\value{
datetime vector
}
\description{
Week datetime from datetime value
}
