% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{increase_datetime_resolution}
\alias{increase_datetime_resolution}
\title{Increase datetime vector resolution}
\usage{
increase_datetime_resolution(y, interval_mins)
}
\arguments{
\item{y}{vector of datetime values}

\item{interval_mins}{integer, interval of minutes between two consecutive datetime values}
}
\value{
datetime vector
}
\description{
Increase datetime vector resolution
}
