% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{increase_timeseries_resolution}
\alias{increase_timeseries_resolution}
\title{Increase time resolution of a timeseries data frame}
\usage{
increase_timeseries_resolution(df, n)
}
\arguments{
\item{df}{data.frame or tibble, first column of name \code{datetime} being of class datetime and rest of columns being numeric}

\item{n}{integer, number of intra-values (counting the original value as the first one)}
}
\value{
tibble
}
\description{
Increase time resolution of a timeseries data frame
}
