% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{interpolation}
\alias{interpolation}
\title{Interpolate \code{n} values between two numeric values}
\usage{
interpolation(y1, y2, n)
}
\arguments{
\item{y1}{first value}

\item{y2}{second value}

\item{n}{integer, number of intra-values (counting the original value as the first one)}
}
\description{
Interpolate \code{n} values between two numeric values
}
