% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpisModule.R
\name{kpisServer}
\alias{kpisServer}
\title{Server function of kpis module}
\usage{
kpisServer(id, kpis_df, kpis_values)
}
\arguments{
\item{id}{character, module ID}

\item{kpis_df}{data.frame or tibble with kpis configuration}

\item{kpis_values}{named list with kpis values}
}
\description{
Server function of kpis module
}
