% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericInputsModule.R
\name{numericInputsUI}
\alias{numericInputsUI}
\title{UI function of numericInput module}
\usage{
numericInputsUI(id, inputs_tbl, sliders = FALSE)
}
\arguments{
\item{id}{character, module ID}

\item{inputs_tbl}{tibble, inputs characteristics}

\item{sliders}{boolean, whether to use sliders or numeric inputs}
}
\value{
shiny UI tagList
}
\description{
UI function of numericInput module
}
