% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_python.R
\name{parse_python_data_frame}
\alias{parse_python_data_frame}
\title{Parse data frame with Python objects to R object}
\usage{
parse_python_data_frame(df)
}
\arguments{
\item{df}{data frame with Python objects}
}
\value{
data frame or tibble
}
\description{
Parse data frame with Python objects to R object
}
