% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_python.R
\name{parse_python_object}
\alias{parse_python_object}
\title{Python to R parsing for DynamoDB items}
\usage{
parse_python_object(object)
}
\arguments{
\item{object}{Python object of class decimal.Decimal, character or logical. For other types a NA value will be returned.}
}
\value{
R object
}
\description{
Python to R parsing for DynamoDB items
}
