% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_python.R
\name{parse_python_objects_list}
\alias{parse_python_objects_list}
\title{Parse a list of Python objects}
\usage{
parse_python_objects_list(objects_list)
}
\arguments{
\item{objects_list}{list of Python objects of class decimal.Decimal, character or logical. For other types a NA value will be returned.}
}
\value{
list of R objects
}
\description{
Parse a list of Python objects
}
