% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_components}
\alias{plot_components}
\title{Function to plot components according to a configuration table}
\usage{
plot_components(df, plot_conf, ylab, legend_width = NULL, css_file = NULL, ...)
}
\arguments{
\item{df}{data.frame or tibble, first column of name \code{datetime} being of class datetime and rest of columns being numeric}

\item{plot_conf}{tibble of the plot configuration with columns \code{variable}, \code{label}, \code{color}, \code{fill} and \code{width} to pass to dygraph.}

\item{ylab}{character, label of y axis}

\item{legend_width}{integer, width (in pixels) of the div which shows the legend.}

\item{css_file}{character path to a CSS file to format dygraph plot. If NULL, custom CSS is applied to dygraph.}

\item{...}{extra arguments to pass to \code{dygraphs::dyOptions} function}
}
\value{
dygraphs plot
}
\description{
Function to plot components according to a configuration table
}
\details{
Attention! By default this function applies dutils CSS file.
}
