% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb.R
\name{put_df_to_dynamodb}
\alias{put_df_to_dynamodb}
\title{Put data frame to DynamoDB table}
\usage{
put_df_to_dynamodb(dynamodb_table, df)
}
\arguments{
\item{dynamodb_table}{boto3 DynamoDB Table object}

\item{df}{data frame to put}
}
\description{
Put data frame to DynamoDB table
}
