% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_paws.R
\name{query_dynamodb_table}
\alias{query_dynamodb_table}
\title{Query items from DynamoDB Table given specific partition and sorting keys}
\usage{
query_dynamodb_table(
  dynamodb_obj,
  table_name,
  partition_key_name,
  partition_key_values,
  sorting_key_name,
  sorting_key_start,
  sorting_key_end,
  parse_item_func = NULL
)
}
\arguments{
\item{dynamodb_obj}{DynamoDB object obtained with \code{get_dynamodb} function}

\item{table_name}{character, DynamoDB table name}

\item{partition_key_name}{character, name of the partiton key of the DynamoDB Table. It has to be a string variable}

\item{partition_key_values}{vector of values of the partiton key of the DynamoDB Table}

\item{sorting_key_name}{character, name of the sorting key of the DynamoDB Table. It has to be a numeric variable}

\item{sorting_key_start}{numeric, start value of the sorting key}

\item{sorting_key_end}{numeric, end value of the sorting key}

\item{parse_item_func}{function to convert every Item from the table to a tibble}
}
\value{
tibble
}
\description{
Query items from DynamoDB Table given specific partition and sorting keys
}
\details{
Note that the amount of items is limited to 1 MB. If the items between the start and end values of the sorting key exceeds this limit,
the data obtained will go from start to the limited maximum.
}
