% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_paws.R
\name{query_dynamodb_table_timeseries}
\alias{query_dynamodb_table_timeseries}
\title{Query table using a timestamp variable as sorting key}
\usage{
query_dynamodb_table_timeseries(
  dynamodb_obj,
  table_name,
  partition_key_name,
  partition_key_values,
  sorting_key_name,
  start_date,
  end_date,
  parse_item_func = NULL,
  tzone = "Europe/Paris",
  query_interval_days = 30,
  milliseconds = TRUE
)
}
\arguments{
\item{dynamodb_obj}{DynamoDB object obtained with \code{get_dynamodb} function}

\item{table_name}{character, DynamoDB table name}

\item{partition_key_name}{character, name of the partiton key of the DynamoDB Table. It has to be a string variable.}

\item{partition_key_values}{vector of values of the partiton key of the DynamoDB Table}

\item{sorting_key_name}{character, name of the sorting key of the DynamoDB Table. It has to be a numeric variable}

\item{start_date}{Date, start date (only class \code{Date}, not \code{datetime})}

\item{end_date}{Date, end date (only class \code{Date}, not \code{datetime})}

\item{parse_item_func}{function to convert every Item from the table to a tibble}

\item{tzone}{character, time zone of the timeseries data}

\item{query_interval_days}{integer, number of days of each query interval}

\item{milliseconds}{logical, whether the timestamp is in milliseconds or seconds}
}
\value{
tibble
}
\description{
Query table using a timestamp variable as sorting key
}
