% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_python.R
\name{query_timeseries_data_table_py}
\alias{query_timeseries_data_table_py}
\title{Query table using a timestamp variable as sorting key}
\usage{
query_timeseries_data_table_py(
  dynamo_table,
  partition_key_name,
  partition_key_values,
  sort_key_name,
  start_date,
  end_date,
  tzone = "Europe/Paris",
  query_interval_days = 30,
  milliseconds = T
)
}
\arguments{
\item{dynamo_table}{boto3 DynamoDB Table obtained with \code{get_dynamo_table} function}

\item{partition_key_name}{character, name of the partiton key of the DynamoDB Table. It has to be a string variable.}

\item{partition_key_values}{vector of values of the partiton key of the DynamoDB Table}

\item{sort_key_name}{character, name of the sorting key of the DynamoDB Table. It has to be a numeric variable.}

\item{start_date}{Date, start date}

\item{end_date}{Date, end date}

\item{tzone}{character, time zone of the timeseries data}

\item{query_interval_days}{integer, number of days of each query interval}

\item{milliseconds}{logical, whether the sorting key is a timestamp in milliseconds or not}
}
\value{
tibble
}
\description{
Query table using a timestamp variable as sorting key
}
