% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb.R
\name{scan_table}
\alias{scan_table}
\title{Scan items from DynamoDB Table given specific numeric attribute range}
\usage{
scan_table(dynamo_table, attribute_name, attribute_start, attribute_end)
}
\arguments{
\item{dynamo_table}{boto3 DynamoDB Table obtained with \code{get_dynamo_table} function}

\item{attribute_name}{character, numeric attribute name}

\item{attribute_start}{numeric, attribute start value}

\item{attribute_end}{numeric, attribute end value}
}
\value{
tibble
}
\description{
Scan items from DynamoDB Table given specific numeric attribute range
}
