% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb_python.R
\name{scan_table_py}
\alias{scan_table_py}
\title{Scan items from DynamoDB Table given specific numeric attribute range}
\usage{
scan_table_py(
  dynamo_table,
  attribute_name,
  attribute_start,
  attribute_end,
  parse = T
)
}
\arguments{
\item{dynamo_table}{boto3 DynamoDB Table obtained with \code{get_dynamo_table} function}

\item{attribute_name}{character, numeric attribute name}

\item{attribute_start}{numeric, attribute start value}

\item{attribute_end}{numeric, attribute end value}

\item{parse}{logical, whether to parse Python objects to R objects}
}
\value{
tibble
}
\description{
Scan items from DynamoDB Table given specific numeric attribute range
}
