% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_detection.R
\name{simulate_step_load}
\alias{simulate_step_load}
\title{Simualte step load}
\usage{
simulate_step_load(time_start, time_end, mins_on, mins_off, power, resolution)
}
\arguments{
\item{time_start}{datetime value to start the simulation}

\item{time_end}{datetime value to finish the simulation}

\item{mins_on}{numeric, number of minutes with step ON}

\item{mins_off}{numeric, number of minutes with step OFF}

\item{power}{numeric, value of power of the step load (in kW)}

\item{resolution}{numeric, minutes interval of the datetime sequence}
}
\value{
data.frame or tibble
}
\description{
Simualte step load
}
