% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{text_info}
\alias{text_info}
\title{Add info icon with a pop-up tip to a label}
\usage{
text_info(label, tip)
}
\arguments{
\item{label}{character, label}

\item{tip}{character, info text}
}
\description{
Add info icon with a pop-up tip to a label
}
