% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{toc}
\alias{toc}
\title{Time difference end function}
\usage{
toc(units = "secs", digits = 2)
}
\arguments{
\item{units}{character, one of "auto", "secs", "mins", "hours", "days" and "weeks"}

\item{digits}{integer, number of decimals}
}
\value{
numeric
}
\description{
Use this function together with \code{tic()} to control time spent by functions
}
