% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericInputsModule.R
\name{update_components_by_inputs}
\alias{update_components_by_inputs}
\title{Update a time-series column according to component input}
\usage{
update_components_by_inputs(df, inputs, prefix, input_factor = 1)
}
\arguments{
\item{df}{data.frame or tibble, being \code{datetime} the first column followed by numeric and UNITARY variables}

\item{inputs}{named list with the name and size of each component}

\item{prefix}{character, prefix of all inputs names}

\item{input_factor}{numeric, factor to apply to each input value when multiplying to the unitary profile.}
}
\value{
tibble, resulting of multiplying unitary vectors by it's corresponding component size
}
\description{
Update a time-series column according to component input
}
\details{
The names of \code{inputs} parameter should correspond to \code{df} variables after splitting the \code{suffix}.
In other words: \code{inputs} = \code{prefix} + \code{names(df)} (of course without considering \code{datetime} column in \code{df})
The returned tibble is the result of \verb{ df` * `inputs` * `prefix}
}
