% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb.R
\name{adapt_date_range}
\alias{adapt_date_range}
\title{Convert start and end dates to DynamoDB timestamp}
\usage{
adapt_date_range(start_date, end_date)
}
\arguments{
\item{start_date}{Date, start date}

\item{end_date}{Date, end date}
}
\value{
tibble
}
\description{
Convert start and end dates to DynamoDB timestamp
}
