% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{adapt_date_range}
\alias{adapt_date_range}
\title{Split the period between start and end dates with a defined interval of days}
\usage{
adapt_date_range(
  start_date,
  end_date,
  tzone = "Europe/Paris",
  interval_days = 30,
  milliseconds = T
)
}
\arguments{
\item{start_date}{Date, start date.}

\item{end_date}{Date, end date}

\item{tzone}{character, time zone of the timeseries data}

\item{interval_days}{integer, number of days of each query interval}

\item{milliseconds}{logical, whether the timestamp variable is in milliseconds or not}
}
\value{
tibble
}
\description{
The dates are converted to timestamp with function \code{date_to_timestamp},
but the inputs \code{start_date} and \code{end_date} must be of class \code{Date} and not \code{datetime}.
}
