% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseSelectorModule.R
\name{caseSelector}
\alias{caseSelector}
\title{Server function of caseSelector module}
\usage{
caseSelector(id, df1, df2, plot_func, plot_conf, ...)
}
\arguments{
\item{id}{character, module ID}

\item{df1}{data.frame or tibble, data of case 1 being \code{datetime} the first column}

\item{df2}{data.frame or tibble, data of case 2 being \code{datetime} the first column}

\item{plot_func}{function to plot the dygraph output, with same parameters as \code{dutils::plot_components} function}

\item{plot_conf}{tibble of the plot configuration to pass to \code{dutils::plot_components} function}

\item{...}{extra parameters passed to dyOptions for \code{df1} and \code{df2} plots (not comparison plot)}
}
\value{
reactive \code{case} module input
}
\description{
Server function of caseSelector module
}
