% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb.R
\name{dynamodb_delete_df}
\alias{dynamodb_delete_df}
\title{Delete data frame from DynamoDB table}
\usage{
dynamodb_delete_df(dynamodb_table, df)
}
\arguments{
\item{dynamodb_table}{boto3 DynamoDB Table object}

\item{df}{data frame to put}
}
\description{
Delete data frame from DynamoDB table
}
