% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamodb.R
\name{dynamodb_put_df}
\alias{dynamodb_put_df}
\title{Put data frame to DynamoDB table}
\usage{
dynamodb_put_df(dynamodb_table, df)
}
\arguments{
\item{dynamodb_table}{boto3 DynamoDB Table object}

\item{df}{data frame to put}
}
\description{
Put data frame to DynamoDB table
}
