% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{format_dygraph}
\alias{format_dygraph}
\title{Format Dygraphs plot with CSS file and extra options}
\usage{
format_dygraph(dyplot, css_file = "www/style.css", ...)
}
\arguments{
\item{dyplot}{dygraph object}

\item{css_file}{character path to a CSS file to format dygraph plot. If NULL, no CSS formating is applied to dygraph.}

\item{...}{extra arguments to pass to \code{dygraphs::dyOptions} function}
}
\value{
dygraphs plot
}
\description{
Format Dygraphs plot with CSS file and extra options
}
\details{
Atention! By default this function needs a CSS file in WWW folder.
}
