% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvgis_api.R
\name{get_multiple_pv_timeseries}
\alias{get_multiple_pv_timeseries}
\title{Get PV generation from PVGIS given the latitude and longitude coordinates and PV installation characteristics.}
\usage{
get_multiple_pv_timeseries(
  year = 2021,
  lat = 52.37,
  lon = 4.908,
  database = "PVGIS-SARAH",
  kWp = 1,
  loss = 14,
  orientation_tbl
)
}
\arguments{
\item{year}{integer, year for the returned time-series data frame}

\item{lat}{numeric, latitude coordinate}

\item{lon}{numeric, longitude coordinate}

\item{database}{character, PVGIS database}

\item{kWp}{numeric, PV installation kWp}

\item{loss}{numeric, PV installation loss}

\item{orientation_tbl}{tibble, with columns \code{varname}, \code{tilt} and \code{azimouth}}
}
\value{
tibble
}
\description{
For more information visit https://ec.europa.eu/jrc/en/pvgis.
For parameters meaning visit https://ec.europa.eu/jrc/en/PVGIS/docs/noninteractive
}
