% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericInputsModule.R
\name{numericInputsServer}
\alias{numericInputsServer}
\title{Server function of numericInput module}
\usage{
numericInputsServer(id, inputs_ids, delay = 0)
}
\arguments{
\item{id}{character, module ID}

\item{inputs_ids}{character vector of inputs to get values from}

\item{delay}{integer, number of milliseconds to delay an input change}
}
\value{
named list with reactive input values
}
\description{
Server function of numericInput module
}
