% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericInputsModule.R
\name{numericInputsUI}
\alias{numericInputsUI}
\title{UI function of numericInput module}
\usage{
numericInputsUI(id)
}
\arguments{
\item{id}{character, module ID}
}
\value{
shiny UI tagList
}
\description{
UI function of numericInput module
}
