% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseSelectorModule.R
\name{plot_components}
\alias{plot_components}
\title{Function to plot components according to a configuration table}
\usage{
plot_components(df, plot_conf, ylab = "kW", ...)
}
\arguments{
\item{df}{data.frame or tibble, first column of name \code{datetime} being of class datetime and rest of columns being numeric}

\item{plot_conf}{tibble of the plot configuration with columns \code{variable}, \code{label}, \code{color}, \code{fill} and \code{width} to pass to dygraph.}

\item{ylab}{character, label of y axis}

\item{...}{extra arguments to pass to \code{dygraphs::dyOptions} function}
}
\value{
dygraphs plot
}
\description{
Function to plot components according to a configuration table
}
