% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericInputsModule.R
\name{update_components_by_inputs}
\alias{update_components_by_inputs}
\title{Update a time-series column according to component input}
\usage{
update_components_by_inputs(df, inputs)
}
\arguments{
\item{df}{data.frame or tibble, being \code{datetime} the first column followed by numeric and UNITARY variables}

\item{inputs}{named list with the name and size of each component}
}
\value{
tibble, resulting of multiplying unitary vectors by it's corresponding component size
}
\description{
Update a time-series column according to component input
}
\details{
The names of \code{inputs} parameter must correspond to \code{df} variables
}
